+++ widgets/barchart/barchartdemo/barchartdemo.go
@@ -25,7 +25,7 @@ import (
 	"github.com/mum4k/termdash/cell"
 	"github.com/mum4k/termdash/container"
 	"github.com/mum4k/termdash/linestyle"
-	"github.com/mum4k/termdash/terminal/termbox"
+	"github.com/mum4k/termdash/terminal/tcell"
 	"github.com/mum4k/termdash/terminal/terminalapi"
 	"github.com/mum4k/termdash/widgets/barchart"
 )
@@ -56,7 +56,7 @@ func playBarChart(ctx context.Context, bc *barchart.BarChart, delay time.Duratio
 }
 
 func main() {
-	t, err := termbox.New()
+	t, err := tcell.New()
 	if err != nil {
 		panic(err)
 	}
@@ -75,10 +75,10 @@ func main() {
 		barchart.ValueColors([]cell.Color{
 			cell.ColorRed,
 			cell.ColorYellow,
-			cell.ColorBlue,
+			cell.ColorNumber(33),
 			cell.ColorGreen,
 			cell.ColorRed,
-			cell.ColorBlue,
+			cell.ColorNumber(33),
 		}),
 		barchart.ShowValues(),
 		barchart.BarWidth(8),